リファクタリング: 重複コードの削除とユーティリティ関数の共通化
- 時間フォーマット関数を utils/timeUtils.ts に共通化
- テキスト切り詰め関数を utils/textUtils.ts に共通化
- Badge コンポーネントを再利用可能なコンポーネントとして抽出
- agents/page.tsx のインポートパスを相対パスに統一
- SessionListView, SessionCard, ConversationCard から重複関数を削除
