+++ src/utils/textUtils.ts
@@ -1,0 +1,4 @@
+export const truncateText = (text: string, maxLength: number = 80): string => {
+  if (text.length <= maxLength) return text
+  return text.substring(0, maxLength) + '...'
+}
